t status git
