+++ Bitcoin6Hourly/.gitignore
