+++ Bitcoin6Hourly/src/main/java/com/jjalgorithms/cryptocurrency/bitcoin/sixhourly/BitcoinScraperApplication.java
@@ -1,0 +1,14 @@
+package com.jjalgorithms.cryptocurrency.bitcoin.sixhourly;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
+
+@SpringBootApplication
+@EnableScheduling
+public class BitcoinScraperApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BitcoinScraperApplication.class, args);
+	}
+}
