+++ Bitcoin6Hourly/src/main/resources/application.properties
@@ -1,0 +1,10 @@
+server.port:9091
+
+#MySQL
+spring.datasource.url=jdbc:mysql://localhost/Bitcoin?autoReconnect=true&useSSL=false
+spring.datasource.username=root
+spring.datasource.password=admin
+spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql = true
