+++ BitcoinDaily/src/main/resources/application.properties
