+++ BitcoinDaily/src/test/java/com/jjalgorithms/cryptocurrency/bitcoin/BitcoinScraperApplicationTests.java
@@ -1,0 +1,16 @@
+package com.jjalgorithms.cryptocurrency.bitcoin;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+public class BitcoinScraperApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
