--- BitcoinScraper/src/main/java/com/jjalgorithms/cryptocurrency/bitcoin/BitcoinScraperApplication.java
@@ -1,14 +1,0 @@
-package com.jjalgorithms.cryptocurrency.bitcoin;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.scheduling.annotation.EnableScheduling;
-
-@SpringBootApplication
-@EnableScheduling
-public class BitcoinScraperApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(BitcoinScraperApplication.class, args);
-	}
-}
